+++ src/events/SubscriptionRequestEvent.php
@@ -7,6 +7,7 @@
 
 namespace craft\commerce\stripe\events;
 
+use craft\commerce\base\Plan;
 use yii\base\Event;
 
 /**
@@ -17,6 +18,11 @@ use yii\base\Event;
  */
 class SubscriptionRequestEvent extends Event
 {
+    /**
+     * @var Plan The subscription plan
+     */
+    public $plan;
+
     /**
      * @var array The subscription parameters
      */